Move plot rmse per bin
